prj-617/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.15.0
google-cloud-storage==2.10.0
pandas==2.0.3
numpy==1.24.3
pandas-gbq==0.19.2
openpyxl==3.1.5
pyarrow==12.0.1